Streamlit_app/requirements.txt
streamlit==1.32.0
faiss-gpu==1.7.4 # Use faiss-gpu if you prefer GPU version
sentence-transformers==2.2.2
langchain-community==0.0.15
torch==2.1.0
numpy==1.24.3
